+++ src/main/java/cz/recipe/dto/RecipeTemplateIngredientDto.java
@@ -13,7 +13,7 @@ import java.io.Serializable;
 public class RecipeTemplateIngredientDto implements Serializable {
 
     private Long id;
-    private float grams;
+    private Long grams;
     private Long ingredientId;
     private Long recipeTemplateId;
